+++ app/src/main/java/de/passwordvault/viewmodel/activities/QualityGatesViewModel.java
@@ -1,0 +1,9 @@
+package de.passwordvault.viewmodel.activities;
+
+import androidx.lifecycle.ViewModel;
+
+public class QualityGatesViewModel extends ViewModel {
+
+
+
+}